converted file to ts
